Update meetupcom lambda in preparation for transformer implementation
