+++ lambdas/meetupcom/package.json
@@ -1,8 +1,7 @@
 {
-  "name": "pull-meetupcom-events",
+  "name": "muxer-pull-events-meetupcom",
   "version": "1.0.0",
   "description": "Pull meetup.com events data",
-  "main": "handler.js",
   "dependencies": {
     "aws-lambda-data-utils": "^1.0.0"
   }
